qwen3_tts_clone_web/requirements.txt
fastapi==0.118.0
python-multipart==0.0.20
python-dotenv==1.1.1
requests==2.32.5
uvicorn[standard]==0.37.0
